+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.7.0
 - New objective handling: Introduce subobjective classes for more intuitive parameterization of weights. As well as for reading out the solved subobjective functions.
 - Update Dashboard to visualize subobjectives
-- Original notation of the objective function still supported. “r_del_u” notation no longer supported.
+- Original notation of the objective function still supported. ``r_del_u`` notation no longer supported.
 
 ## 0.6.8
 - #55 improved numerics of casadi model simulation by adding initial guess for outputs